dir improvements: move `eth2shuffle`from `erigon-lib` to `cl/utils`
