+++ cl/utils/eth2shuffle/shuffle_bench_test.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/erigontech/erigon-lib/common/eth2shuffle"
+	"github.com/erigontech/erigon/cl/utils/eth2shuffle"
 )
 
 func BenchmarkPermuteIndex(b *testing.B) {
